feat: ListItem 변경점 story에 적용
